CLN : Update super usage
On Python 3, super can be called without any arguments where as on
Python 2, super needs to be called with the class object and an
instance.

Given that dask does not support Python 2 anymore, the super usage can
be updated.

Note that this was an automated regex-based search and replace. After
the automated search and replace happened, each change was individually
checked and committed.

	modified:   dask/dataframe/core.py
	modified:   dask/dataframe/groupby.py
	modified:   dask/delayed.py
	modified:   dask/diagnostics/profile.py