Partially revert chunk-splitting in indexing.
This partially reverts the changes made in
https://github.com/dask/dask/pull/6514. It restores the old behavior
with a warning that large chunks (10x array.chunk-size) are being
produced.

Additionally, it adds a new config option to control the behavior
(`array.slicing.split-large-chunks`). Setting that to `False` silences
the warnings and keeps the "old" behavior (one output block per input
block touched, even if this makes a large output). Setting that to
`True` silences the warning and restores the Dask 2.26 behavior of
splitting.

Closes https://github.com/dask/dask/issues/6646