Correctly tear down LoopRunner in Client (#4112)
See #4072

LoopRunner is designed to stop loops it started.
`loop` fixture doesn't start its IOLoop but always stops it.
This leads to conflict in the test: both LoopRunner and `loop` fixture
stop the same loop. So we need to use `loop_in_thread` fixture
since it starts and stops its IOLoop.

`_should_close_loop` flag is not used, moreover, LoopRunner detects
already started loop in the `_start_unlocked()` and doesn't stop it
in the `_stop_unlocked()`.