Improve ordering for workloads with a common terminus (#6779)
The current ordering algorithm struggles with workloads where there's
a single, common root node that's downstream of a tasks with some shared
dependencies.

                s
        /   /       \   \
       /   /         \   \
     s00  s10       s01  s11
      |    |         |    |
     b00  b10       b01  b11
     / \  / \       / \ / \
    a0  a1  a2    a3  a4  a5

Previously, we started at a0, finished s00, and then jumped down to a3
after deciding that we should work on s01 based on a tiebreaker. We'd
instead like to work on s10, thakns to the shared dependency a1.

The ordering algorithm does just fine in the absence of the common root
node `s`. So this PR proposes to just ignore it, at least in the part of
the ordering where we made the wrong choice previously.