Add inline_array option to from_array (#6773)
This adds an option to `da.from_array` to control how the array object
itself is included in the task graph. By default, there's no change:
there will be a single key whose value is the array object. The tasks
giving each chunk of the array refer to the array by using its key.

With `inline_array=True`, each chunk task in the task graph will have
the array itself included in the values.