HighLevelGraph length without materializing layers
Closes #7271

Calculate the `__len__` of a HighLevelGraph from the sum of the legths of its layers, instead of `len(self.to_dict())`. This is much faster and prevents causing all the layers to materialize.

I also changed the `__len__` implementation on `Blockwise`. It was using `_out_numblocks`, which is unused anywhere else in the codebase, and a bit hard to read. As far as I could tell, `_out_numblocks` was equivalent to `{i: self.dims[i] for i in self.output_indices}`. Basically, the length of a Blockwise layer should be equal to the number of output keys (right?).