Check that computed chunks have right size and dtype
This was inspired by a question in #7234. At present `assert_eq` checks
that the computed array has the right size and dtype, but it's still
possible that the individual chunks have the wrong size (as long as they
add up to the right total size) and maybe the wrong dtype (as long as
they coerce to produce the right dtype - although I haven't actually
checked that this would slip through).

Now `persist` is used to access the individual computed chunks and
checked for shape and dtype.

This is a draft because there are a couple of failures. One is in
`test_zarr_return_stored[False]` (which I'll need to investigate
further, but possibly `persist` just doesn't make sense for this case)
and the other is in `test_blockwise_concatenate` which looks like it
might just be a test bug.