Check that computed chunks have right size and dtype (#7277)
* Check that computed chunks have right size and dtype

This was inspired by a question in #7234. At present `assert_eq` checks
that the computed array has the right size and dtype, but it's still
possible that the individual chunks have the wrong size (as long as they
add up to the right total size) and maybe the wrong dtype (as long as
they coerce to produce the right dtype - although I haven't actually
checked that this would slip through).

Now `persist` is used to access the individual computed chunks and
checked for shape and dtype.

* Fix test_blockwise_concatenate

It had a `blockwise` function that returned a shape inconsistent with
that inferred from the input block sizes.

* Fix dtype handling for *_like_safe on old numpy

It was being used with non-arrays, so explicitly force things to be
arrays.