Batch multiple tasks with `executor.submit`
Estimates a rough `chunksize` that evenly divides the work between each
batch submission and submits it.

In theory `Executor.map` should be used for this. However that runs into
issues on Windows where problems with the `Queue` not yielding to
interrupts would hang things for the user. Plus `map` is synchronous and
blocks for all work to complete, which we don't want either. By using
`executor.submit` and batching ourselves we are able to avoid both of
these issues.