Support out-of-band pickling in multiprocessing
Tweak the implementations of `_dumps` and `_loads` to optionally handle
out-of-band buffers during pickling. This can help avoid unnecessary
copying of data before sending it over the wire. We do this by storing
the pickled result in a `list` instead of a `bytes` object. The first
element is what `dumps` would produce or `loads` would consume, but the
rest of the elements are buffers that pickling has extracted and sent
alongside. This analogous to the strategy we use in Distributed. As a
result large array or dataframes, can be sent along without copying
their content necessarily. Thus putting the multiprocessing use case on
more even footing with the threaded or distributed use cases.

Note that if pickle protocol 5 is not supported, we just send over a
singleton `list` bypassing all out-of-band buffer handling. As
cloudpickle can detect whether pickle protocol 5 is supported, we use
that determine whether to leverage the out-of-band pickling case or not.