Put batched results as one item in the queue
Instead of putting each result from a batched submission in the queue
separately, just put it as one item in the queue. This avoids multiple
call to `put` and possible deadlocks this may entail. After pulling the
batched results out of the queue, iterate over them in the main thread
until they are all handled.