Use `for`-loop to iterate over `"ready"` keys
Instead of `pop`ing each `key` in `"ready"` within a `while`-loop, just
iterate over the `key`s in `"ready"` with a `for`-loop and `clear` the
`list` once that is done. This should avoid resizing the `list` multiple
times and repeated checking of its content.

Note that as we were `pop`ing from the end, we need to have the `list`
`reversed` first to iterate through it in the intended order.