Make `chunksize` configurable (default to `1`)
As the appropriate `chunksize` for workloads can vary depending on the
type of workload, make this configurable instead of simply picking a
value. `get_async` will accept this as an argument. Though if
`chunksize` is not provided, `get_async` will check the `dask.config`.
By default set this to `1`, which matches the local schedulers original
behavior and what `map` does in `concurrent.futures`. To evenly divide
work across workers, this can be set to `-1` and will be computed
under-the-hood (the behavior up until now). Any other `int` will be used
to pick a suitable `chunksize`.