Only provide installation instructions when distributed is missing
The line "from dask.distributed import ..." causes us to import
distributed.
If that doesn't work today, we currently tell people to go and install
distributed with pip/conda.

However, if there is an underlying optional library that is missing
(like uvloop)
then this message hides the actual issue and so leads the user astray.

This commit restricts the use of the informative message to only the
case when distributed is missing, and falls back to normal Python
behavior otherwise.