histogram, histogramdd improvements (docs; return consistencies)
* Expand the docstring of dask.array.histogram such that it's clear
why a Series can sometimes be passed to the function (and when it
can't be passed).

* Improve the docstring of dask.array.histogramdd and adjust the
second return (the bin edges) as well (originally a list of
numpy.ndarrays was returned, but for consistency with
dask.array.histogram we change it to a list of dask.array.Arrays). Add
a test for the second return.