Generate unique names for SubgraphCallable
If no name is supplied by the caller, use `tokenize` to generate a name
rather than just calling it "subgraph_callable". This ensures that
(anonymous) graphs with the same outkey and inkeys (which for Blockwise
really only captured the arity since placeholder inkeys are used) but
different graphs (due to subgraph fusion) do not compare equal.

Closes #7632.