Switch SubgraphCallable.__eq__ to use UUID
This means instances that are otherwise identical but were not
explicitly named will compare unequal, which is a tradeoff to allow
construction to be faster.

The unit tests for SubgraphCallable need a tweak to use the old method,
so that instances created by the code under test can compare equal to
the reference values.