Implement map_blocks with block_info via BlockwiseDep
This still assembles the block_info parameter when the Blockwise layer
is materialised, rather than as part of the task, so it is probably not
going to be any more scalable, but it avoids creating an additional
layer.

It still needs to be updated with serialization support to make it work
with distributed.