Rewrite BlockwiseDep handling of map_blocks
This is an attempt to improve the serialization situation. Instead of
trying to put the data into a custom subclass of BlockwiseDep, insert it
as (constant) arguments to the wrapper function. This then relies on
SubgraphCallable to handle the serialization.

This may theoretically improve serialization costs when the graph is
materialized on the client (which I think is still the default for
arrays), because the raw data is inside the SubgraphCallable and hence
only serialised once, with production of the individual block_infos left
to the workers.

The benchmark code in #7686 is a bit slower than the previous version
(about 7s for compute), but still faster than main.