Do not return delayed object from to_zarr
The change to wrap zarr.create in dask.delayed was originally made by me in #5797.  I did this when I was still new to dask, and when it made sense (to me) to defer any file operations, including the initial metadata creation, until run time.

After having worked with the dask ecosystem for a while, I see that this is not the typical paradign for I/O.  A call to to_zarr should immediately create the zarr metadata, so as not to surprise others.