Automatically show graph visualization in jupyter notebooks (#7716)
I noticed that `result.dask.visualize()`  / `result.__dask_graph__().visualize()` in a jupyter notebook, it writes the graph visualization to file, but *doesn't* automatically show you the graph in the notebook cell output.

In contrast, when you call `result.visualize()` on a dask computation in a jupyter notebook, you are automatically shown the visualization of the low level task graph in the notebook cell output. This is convenient, and I'd like both visualize functions to have this behaviour.

This PR changes it so we do automatically see the high level graphs visualized in the jupyter notebook.