Improvements to histogramdd (for handling inputs that are sequences-of-arrays). (#7634)
This PR improves `dask.array.histogramdd` such that multi-sequence input does not have to be stacked, transposed, and potentially rechunked.

The original implementation of `histogramdd` was focused on columnar/rectangular input, with each column of the input array representing one coordinate of the input to be histogrammed. If the input was a sequence of 1D arrays then we stacked the 1D arrays and transposed that stack to convert to the 2D/rectangular/columnar format we were mainly supporting. This stacking/fusing and transposing is an unnecessary step.

The rectangular input implementation is still there, but now we handle the sequence-of-1D inputs without converting to a 2D array. This optimization is important for supporting `numpy.histogram2d` as `dask.array.histogram2d` (subsequent PR).

Along with supporting sequence-of-1D input, this PR also does some general cleanup.