Add colors to represent high level layer types (#7974)
Added an extra attribute to the graphviz output of High-Level graphs - node fill color. Nodes are colored on the basis of their layer_type.

This was achieved by using DOT's fillcolor attribute.

Currently, this option is made optional to the users. They need to provide an additional kwarg while calling dask.visualize() with color="layer_type".