A few threaded scheduler fixups
Previously there was a bug where the number of threads used by the
threaded scheduler was different if `dask.compute` was called in the
main thread vs in a background thread. While fixing this, I also noticed
that the logic around maintaining a cache of executors could be done in
a simpler way that avoided the need for a lock or active management (we
could instead rely on GC behavior). This patch includes fixes for both.