Dispatch walks mro for lazily registered handlers (#8185)
Previously when `Dispatch` walked the `mro` of a class, it would only
look for eagerly registered handlers (instead of lazily registered
handlers). This meant that subclasses of lazily registered handlers
would only be used when dispatching on a subclass if the parent class
had been dispatched on earlier (this triggering the registration).

This patch fixes this to check for lazily registered handlers when
walking the mro. This results in negligible slowdown during the fast
path.

Fixes #8111.