Ensure existing futures get passed as deps
A bug in the serialization of HighLevelGraph layers resulted in tasks in
a layer that depend on an existing future not having that future marked
as a dependency for the task (resulting in the future name being passed
instead of its value). This took forever to debug.

Rearranging the merge order here so known futures don't get overwritten
fixes things. I've also taken the opportunity to clean up this code a
bit and add some comments. No tests yet.