Support negative drop_axis in map_blocks and map_overlap (#8192)
This PR allows the `drop_axis` argument of `map_blocks` and `map_overlap` to use negative axis values in the same manner as functions in the NumPy API. 

This improves on #8064 which added a more informative error message on negative `drop_axis` (closing #7924). This PR changes the error to only occur when the specified axis exceeds the dimensions of the output array, otherwise the axis is converted to the equivalent positive-valued axis.