Fix test_describe_empty to work without global -Werror
Fix test_describe_empty to work when the test suite is run without
global -Werror.  This is e.g. desirable for packagers who don't want
the test suite for a fixed version to suddenly start failing due to
DeprecationWarnings in dependencies that otherwise don't break
the package.

Since the test expects either a ValueError or a RuntimeWarning, it seems
that the easiest way to assert for that is to inject the "error" filter
for the scope of the call.