Add `align_dataframes` to `map_partitions` to broadcast a dataframe passed as an arg (#6628)
When dataframes are passed to `map_partitions`, by default they are aligned according to the partitions of the first dataframe. This can be surprising when the user intended to broadcast them instead. This PR adds `align_dataframes` as a `map_partitions` kwarg (just like `align_arrays` in `map_blocks`) so you can work around this when necessary (without going the hacky route of using kwargs when you intend to broadcast).

Co-authored-by: Gabe Joseph <gjoseph92@gmail.com>