Get tests working on p2p shuffle
All tests pass except `test_shuffle_sort`, because it has the gall to slice from the DataFrame after shuffling! This causes rows to get culled, so some of the output tasks we try to set worker restrictions on don't exist.

For how simple it is, this is a very broken case in the overall shuffle design. It's completely dependent on none of the output tasks getting culled. And it's actually not so simple to solve, because
a) we should do the shuffle differently if we don't need all output partitions
b) figuring out which tasks have been culled is tricky; requires either custom optimization or talking to scheduler at runtime
c) the outputs needed _could change_ at runtime (first you compute just 5 partitions, then immediately change your mind and submit the whole thing—the `transfer`/`barrier` tasks haven't changed, nor have the `unpacks`. There are just more `unpacks`. But the new `unpacks` are actually broken if we cleverly skipped sending some data in the original `transfer`s.)