Allow custom sort functions for `sort_values` (#8345)
Allows the sorting function called on each partition in last step of `sort_values` to be generalized, along with the kwargs that are supplied to it. This allows `sort_values` to be extended to support multi-column sorting and more complex ascending / null position handling.

The context for this PR is a desire to simplify the [sorting algorithm](https://github.com/dask-contrib/dask-sql/blob/main/dask_sql/physical/utils/sort.py) used by dask-sql; since it only really differs from Dask's sorting algorithm in that it uses a custom sorting function, it seems like it would be easier to allow for that extension upstream rather than duplicate Dask code in dask-sql.