Improved the efficiency of `matmul()` by completely removing concatenation (#8423)
This PR avoids concatenation in both `blockwise()` and `reduction()` within the implementation of `matmul()`.  Previously, `matmul()` avoided concatenation only in `blockwise()`.

Due to the special nature of the preceding blockwise contraction, all chunk-operands of a sum during the reduction phase are expected to have exactly the same shape, with a size of 1 for the reduction axis.  This makes mere element-wise addition of the chunks possible.   Furthermore, the summation result can be merely reshaped to lose the reduction axis when requested.  In principle, both these operations (element-wise addition and reshaping) are performance boosters compared to the default concatenation-mediated reduction which often involves data rearrangement.

The code of `matmul()` has also been shortened, making it a bit more readable.  This is a direct consequence of ensuring that the position of the new-axis introduced in the output of chunk-function `_matmul()` \[the first argument to `blockwise()`\] is consistent with its expected position in the output of `blockwise()` (the position of the contraction-axis).

Performance is comparable to the original `matmul()` implementation in all cases except one, in which this new implementation is approximately 20% faster, namely, "the case where all dimensions are chunked".  This outcome is consistent with the total absence of concatenation-overhead in the new version.   Performance results can be found in this [jupyter notebook](https://github.com/ParticularMiner/dask/blob/matmul_performance/matmul_perf_second2last.ipynb).

Co-authored-by: Peter Andreas Entschev <peter@entschev.com>