TST: fix parquet test_pandas_timestamp_overflow_pyarrow test (#8733)
This test started failing on arrow master (noticed thanks to our nightly builds testing dask). The reason for the failure is that we started using the `types_mapper` keyword in ``pa.Array.to_pandas`` (previously this was ignored for arrays), but in this test it is passed a wrong values (it should be a callable, not a dict).