Finish making ``map_overlap`` default boundary ``kwarg`` ``'none'`` (#8743)
Followup to PR https://github.com/dask/dask/pull/8397

We've had a FutureWarning up for a few months about an upcoming change to the default 'boundary' kwarg value in `map_overlap`, so now is the time to change it. Previous default was `"reflect"`, new default will be "None".

The reason for this change is that it makes the code run a lot faster, and for most people the overlap depth is sufficient and they should not require additional boundary handling. See https://github.com/dask/dask/issues/8391 for a full discussion.