Add sanity checks to divisions setter
Adds a few sanity checks to the divisions property setter

- New divisions are compatible with the existing `npartitions`
- New divisions don't mix None and non-None values
- New divisions are sorted (except for ordered categorical dtypes, where
that's hard)
- New divisions are unique except for the last value

These sanity checks turned up a number of bugs that needed to be fixed.
- `map_partitions` with `align_dataframes=False` would select the wrong
output divisions, any use of this was relying on the caller manually
fixing things later.
- `set_index` could result in non-unique divisions being used in certain
cases (e.g. `(0, 0, 0, 1, 2, 3, 3)`). We now deduplicate values in
places where divisions are computed as needed.
- A few category-dtype related fixes. Category dtypes are leaky
abstractions, all of this feels a little hacky, but tests seem to pass.