to_parquet default to `compression="snappy"`
Previously `to_parquet` had `compression="default"` as the default. In
this case, `"snappy"` was used if `import snappy` succeeded, otherwise
`None` was used.

This may have been correct when written, but is no longer correct.

- `fastparquet` uses `cramjam` for compression support (changed ~1 year
   ago). This library is not optional, so `fastparquet` always supports
   snappy.
- `pyarrow` uses the snappy c library directly (and it's always
  included), whether `snappy` (the python library) successfully imports
  is irrelevant.

As such we chose to deprecate `compression="default"` in favor of using
`"snappy"` everywhere for consistency across engines.