Use Blockwise/`map_partitions` for more tasks (#8831)
* hash_join uses map_partitions

This replaces the low-level graph generation with map_partitions to
generate the high-level graph. Discussed in GH8306.

* stack_partitions uses HLG

Creates a high-level graph wrapper for the partition stacking operation.
Removing the forced materialization of the incoming DataFrame means that
any blockwise operations on the DataFrame will remain HLG layers at this
point.

* Refactors stack_partitions concat tests

Several tests appeared to be duplicated from the first for-loop in
test_concat. These are removed. The remaining tests are moved to a new
function since they test Series concat. These tests are also
de-duplicated.

* Adds an HLG-assurance test

With the changes here, blockwise operations should remain blockwise for
concat operations.