Requires parquet in predicate to be iterable
As reported in #8720, the _flatten_filters function required the value
to be hashable. This implicitly required the value to be a tuple,
although lists and sets would also be appropriate since they support the
'in' operation. _flatten_filters was only used to determine which
columns should be filtered, so it can be refactored and removed.

The 'in' predicate also requires that the value be iterable.
Non-iterable values are now caught and an appropriate message is raised
indicating what the user needs to change. This must be done in two
places to support both fastparquet and pyarrow.