Only test on Python 3.9 on macos (#9029)
The macos builds consistently take the longest to run, and have a lower
org-level concurrency limit than windows or linux. In `dask/dask` we
have no logic specific to macos, so testing on every python version on
this platform doesn't seem necessary. To reduce CI delay, we change
github actions to only test macos on python 3.9.