Prevent codecov from notifying of failure too soon
Currently codecov pushes an initial failing status on every PR that is
later updated to passing once more CI builds finish. This is annoying.

As far as I can tell from https://docs.codecov.com/docs/merging-reports
this shouldn't be happening, but it is. Here we attempt to stop this
behavior by:

- Only running code coverage on the main test suite that hits most of
  the codebase.
- Setting a minimum number of builds before codecov should push a status
  update.

Hopefully this fixes the problem.