Patch around pandas ArrowStringArray pickling (#9024)
The (experimental) pandas `string[pyarrow]` dtype has some major
performance benefits that we'd like to experiment with in dask. However,
currently `pyarrow.StringArray` objects have a bug in their pickle
implementation where a small slice of the array still serializes the
full (potentially very large) backing buffers (see
https://issues.apache.org/jira/browse/ARROW-10739). Hopefully this is
fixed upstream in pyarrow at some point, but for now we patch around it
by overriding the pickling implementation for `ArrowStringArray` in
pandas. This implementation is efficient, resulting in zero-copy
serialization in most cases.

There is still more work to do to fully support the `string[pyarrow]`
dtype, but I think this PR can go in as is for now.