Remove ensure_file from recommended practice for downstream libraries
Previously we did this to educate users about possible configuration.
But as configuration shifts really all this means is that there's lots
of stale config files lying around.

Historically we've found that it's better not to copy these.
We've been doing this in Dask libraries for some time now.
This change just changes the recommendations in the documentation.