Dask in pyodide (#9053)
This is a small step towards #7764 and https://github.com/dask/distributed/issues/6257. It's basically just defensively importing `threading` and `multiprocessing` and defaulting to the synchronous scheduler if those fail. So this is currently mostly be useful for demos and training around the dask collections API. But it *does* work.

This is distinct from actually getting a `distributed.Client` working and talking to a remote cluster, which will require some actual networking work.