Make cycle reported by toposort shorter. (#9068)
This is an alternative to #9057.  Previously, the cycle reported by toposort could contain extra dependencies of nodes that are part of the cycle.  Now we use a simple heuristic to greedily construct a short cycle that only contains nodes in the cycle, but there is no guarantee that this is the shortest cycle.