Update deploying-python.rst
Since memory_limit is not part of the member args, I took a look at the sphinx documentation and found that :undoc-members: generates a doc for members not having docstrings (which I believe there is no docstring for memory_limit). Please let me know if I'm on the right track - still fairly new to sphinx as well as dask but eager to work on this issue!