Get kinda working with core blockwise
Graph isn't computable, fails during optimization. The problem is that because we pre-optimize all the inputs together, their names don't exist anymore in `dependencies`. So something in `fuse_roots` is breaking looking up the dependency of one of our blockwise layers, which doesn't actually exist. (An odd case where Blockwise refers to a name which _does_ exist in the materialized graph, but doesn't exist in the layers.)