Switch default bitgenerator from PCG64 to PCG64DXSM
Uses of the PCG64 BitGenerator in a massively-parallel context have been
shown to have statistical weaknesses that were not apparent at the first
release in numpy 1.17. Most users will never observe this weakness and
are safe to continue to use PCG64. We have introduced a new PCG64DXSM
BitGenerator that will eventually become the new default BitGenerator
implementation used by default_rng in future releases. PCG64DXSM solves
the statistical weakness while preserving the performance and the
features of PCG64.

https://numpy.org/doc/stable/reference/random/upgrading-pcg64.html