limit max chain traversal with each recursion
This is meant to prevent over-eating tree reductions, and it seems to work! `test_two_step_reduction` passes now. But it's broken if you add a widely-shared dep. I think the concept is right but the specific logic is off. `test_vorticity` is broken now. The linear case passes, but seems off (too few groups). Others are missing the second-order grouping we want.