Rename columns correctly in case of SeriesGroupby
When aggregating on a SeriesGroupby, the column we handle is only a
single string (rather than a list of column names), so we must be
careful not to explode into a set of characters when checking if we
should internally rename the grouping column.

While we're here, only rename those columns that need it.

- Closes #9715.
- Closes #9313.