POC - Add basic shim layer class for expressions
This is for demonstration purposes only.

This experiment introduces a non-functional expression layer in between
the DataFrame/Series classes and meta/divisions/name/graph generation.
This expression layer doesn't do anything, in only holds a single opaque
type, but it demonstrates the ease/pain with which we could introduce a
new abstraction and slowly phase out immediate metadata construction.

Most of the pain here was about being careful with attribute assignment
/ properties / getters and setters.  I didn't bother with all of this so
about half of the test suite fails.  This is enough though, I think, to
show a path.