Relax assert_eq checks on Scalar-like objects
This is part of an effort to reuse assert_eq on other objects that are
similar to Dask Dataframe, but not exactly Dask dataframe objects.

This was done previously with Series/Dataframe types above,
now we allow any object that isn't a Series/DataFrame/Index to be
interpretted as a Scalar.  Previously this would raise an exception.