Handle None column name when checking if columns are all numeric
We cannot ask for the names of the grouper since if the data frame has
a column named None we cannot distinguish it as a grouping column from
the case where the groupby argument is not a column selection (e.g. a
function that assigns rows to groups). To fix this, just perform the
groupby on meta and see what we get out.

Closes #10127.