Move _task up into Expr (#30)
Previously expressions defined _layer methods.  Most of them were of the
form

```python
return {(self._name, i): f(i) for i in range(self.npartitions)}
```

We took advantage of this structure for Blockwise Fused expressions, but
maybe it's a good general default to have more often.  This moves the
generic version of this up from Blockwise to Expr, and reuses the
structure in Head and Partitions, leaving only FromGraph and
ApplyConcatApply with custom _layer definitions.