Avoid splitting parquet files to row groups as aggressively
This makes it less likely that we'll split up parquet files.  It does so
in two changes:

1.  Change the default blocksize from 128 MIB to 256 MiB
2.  Only trigger this check if we double the optimal blocksize (512 MiB)

    (but then presumably we'll target 256)

Additionally, for follow-up work it would be excellent to make this
sensitive to only reading certain columns.  Right now I think we do this
even if we read only a single column.

There's plenty more that can be done here, but hopefully this is easy to
agree on as a first step.